Split parser into pieces
