+++ packages/parser/src/index-v06.ts
@@ -1,0 +1,2 @@
+export { toXml } from "./dast-to-xml/dast-util-to-xml";
+export { lezerToDastV6 } from "./lezer-to-dast/lezer-to-dast-v6";
