+++ packages/parser/src/index.ts
@@ -6,7 +6,5 @@ export * from "./detect-version";
 export { visit } from "./pretty-printer/normalize/utils/visit";
 export { toXml } from "./dast-to-xml/dast-util-to-xml";
 export { lezerToDast, stringToLezer } from "./lezer-to-dast/lezer-to-dast";
-export { lezerToDastV6 } from "./lezer-to-dast";
-export { prettyPrint } from "./pretty-printer";
 export { filterPositionInfo } from "./dast-to-xml/utils";
 export { normalizeDocumentDast } from "./dast-normalize/normalize-dast";
