+++ packages/parser/src/lezer-to-dast/index.ts
@@ -1,2 +1,1 @@
 export { lezerToDast } from "./lezer-to-dast";
-export { lezerToDastV6 } from "./lezer-to-dast-v6";
