+++ packages/parser/vite.config.ts
@@ -24,16 +24,10 @@ export default defineConfig({
         lib: {
             entry: {
                 index: "./src/index.ts",
+                "pretty-printer": "./src/pretty-printer/index.ts",
+                "index-v06": "./src/index-v06.ts",
             },
             formats: ["es"],
-        },
-        rollupOptions: {
-            // This is needed for building the `v06-to-v07` converter.
-            // We don't want to add a build dep on `@doenet/doenetml-worker-javascript` just for this.
-            external: [
-                "@doenet/doenetml-worker-javascript",
-                "lib-doenetml-worker",
-            ],
         },
     },
 });
